test: migrate webview attribute specs to spec-main
